Allow workload resolvers to use workload installation record library
