+++ src/Cli/dotnet/commands/dotnet-workload/install/IWorkloadPackInstaller.cs
@@ -3,6 +3,7 @@
 
 using System.Collections.Generic;
 using static Microsoft.NET.Sdk.WorkloadManifestReader.WorkloadResolver;
+using Microsoft.DotNet.Workloads.Workload.Install.InstallRecord;
 
 namespace Microsoft.DotNet.Workloads.Workload.Install
 {
