+++ src/Cli/dotnet/commands/dotnet-workload/install/WorkloadInstallCommand.cs
@@ -17,6 +17,7 @@ using Microsoft.DotNet.Configurer;
 using NuGet.Protocol;
 using NuGet.Protocol.Core.Types;
 using NuGet.Versioning;
+using Microsoft.DotNet.Workloads.Workload.Install.InstallRecord;
 
 namespace Microsoft.DotNet.Workloads.Workload.Install
 {
