+++ src/Cli/dotnet/commands/dotnet-workload/install/WorkloadInstallRecords/IWorkloadInstallationRecordInstaller.cs
@@ -1,0 +1,18 @@
+// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Collections.Generic;
+
+namespace Microsoft.DotNet.Workloads.Workload.Install.InstallRecord
+{
+    internal interface IWorkloadInstallationRecordInstaller
+    {
+        IEnumerable<string> GetInstalledWorkloads(SdkFeatureBand sdkFeatureBand);
+
+        void WriteWorkloadInstallationRecord(WorkloadId workloadId, SdkFeatureBand sdkFeatureBand);
+
+        void DeleteWorkloadInstallationRecord(WorkloadId workloadId, SdkFeatureBand sdkFeatureBand);
+
+        IEnumerable<SdkFeatureBand> GetFeatureBandsWithInstallationRecords();
+    }
+}
