+++ src/Cli/dotnet/commands/dotnet-workload/install/WorkloadInstallRecords/NetSdkManagedInstallationRecordInstaller.cs
@@ -1,0 +1,70 @@
+// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+
+namespace Microsoft.DotNet.Workloads.Workload.Install.InstallRecord
+{
+    internal class NetSdkManagedInstallationRecordInstaller
+    {
+        private readonly string _workloadMetadataDir;
+        private readonly string _installedWorkloadDir = "InstalledWorkloads";
+
+        public NetSdkManagedInstallationRecordInstaller(
+            string dotnetDir)
+        {
+            _workloadMetadataDir = Path.Combine(dotnetDir, "metadata", "workloads");
+        }
+
+        public IEnumerable<SdkFeatureBand> GetFeatureBandsWithInstallationRecords()
+        {
+            if (Directory.Exists(_workloadMetadataDir))
+            {
+                var bands = Directory.EnumerateDirectories(_workloadMetadataDir);
+                return bands
+                    .Where(band => Directory.Exists(Path.Combine(band, _installedWorkloadDir)) && Directory.GetFiles(Path.Combine(band, _installedWorkloadDir)).Any())
+                    .Select(path => new SdkFeatureBand(Path.GetFileName(path)));
+            }
+            else
+            {
+                return new List<SdkFeatureBand>();
+            }
+        }
+
+        public IEnumerable<string> GetInstalledWorkloads(SdkFeatureBand featureBand)
+        {
+            var path = Path.Combine(_workloadMetadataDir, featureBand.ToString(), _installedWorkloadDir);
+            if (Directory.Exists(path))
+            {
+                return Directory.EnumerateFiles(path)
+                    .Select(file => Path.GetFileName(file));
+            }
+            else
+            {
+                return new List<string>();
+            }
+        }
+
+        public void WriteWorkloadInstallationRecord(WorkloadId workloadId, SdkFeatureBand featureBand)
+        {
+            var path = Path.Combine(_workloadMetadataDir, featureBand.ToString(), _installedWorkloadDir, workloadId.ToString());
+            var pathDir = Path.GetDirectoryName(path);
+            if (pathDir != null && !Directory.Exists(pathDir))
+            {
+                Directory.CreateDirectory(pathDir);
+            }
+            File.Create(path);
+        }
+
+        public void DeleteWorkloadInstallationRecord(WorkloadId workloadId, SdkFeatureBand featureBand)
+        {
+            var path = Path.Combine(_workloadMetadataDir, featureBand.ToString(), _installedWorkloadDir, workloadId.ToString());
+            if (File.Exists(path))
+            {
+                File.Delete(path);
+            }
+        }
+    }
+}
