+++ src/Tests/dotnet-workload-install.Tests/GivenNetSdkManagedWorkloadInstall.cs
@@ -17,6 +17,7 @@ using NuGet.Versioning;
 using Xunit;
 using Xunit.Abstractions;
 using static Microsoft.NET.Sdk.WorkloadManifestReader.WorkloadResolver;
+using Microsoft.DotNet.Workloads.Workload.Install.InstallRecord;
 
 namespace Microsoft.DotNet.Cli.Workload.Install.Tests
 {
